Added typing command to chat.session.
Also added typing function and member_typing event to nkchat_conversation.
When a user sends a typing command to an active conversation, a "member_typing" event will be broadcasted to all chat.sessions in that conversation.
Besides, the user who call the typing command will be marked as "active" (that is, with a green badge in the MPS client).