Return cached member data instead of object data.
This will fix a bug when deactivating a conversation and reactivating the same conversation right after sending a new message. Now, messages that haven't been saved yet to the object will be showed with the correct check mark (double blue check if the message was read by all users in that conversation, etc.).